# python-flask/swagger_server/models/command.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Command(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, text: str=None, cmd: str=None):  # noqa: E501
        """Command - a model defined in Swagger

        :param text: The text of this Command.  # noqa: E501
        :type text: str
        :param cmd: The cmd of this Command.  # noqa: E501
        :type cmd: str
        """
        self.swagger_types = {
            'text': str,
            'cmd': str
        }

        self.attribute_map = {
            'text': 'text',
            'cmd': 'cmd'
        }
        self._text = text
        self._cmd = cmd

    @classmethod
    def from_dict(cls, dikt) -> 'Command':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Command of this Command.  # noqa: E501
        :rtype: Command
        """
        return util.deserialize_model(dikt, cls)

    @property
    def text(self) -> str:
        """Gets the text of this Command.


        :return: The text of this Command.
        :rtype: str
        """
        return self._text

    @text.setter
    def text(self, text: str):
        """Sets the text of this Command.


        :param text: The text of this Command.
        :type text: str
        """
        if text is None:
            raise ValueError("Invalid value for `text`, must not be `None`")  # noqa: E501

        self._text = text

    @property
    def cmd(self) -> str:
        """Gets the cmd of this Command.


        :return: The cmd of this Command.
        :rtype: str
        """
        return self._cmd

    @cmd.setter
    def cmd(self, cmd: str):
        """Sets the cmd of this Command.


        :param cmd: The cmd of this Command.
        :type cmd: str
        """
        if cmd is None:
            raise ValueError("Invalid value for `cmd`, must not be `None`")  # noqa: E501

        self._cmd = cmd